Remove accidental Guava dependency
Guava was an implicit dependency through libmtg, until I removed it from
there -- and realized I was using it in this project without declaring
it as a dependency. Oops!

Rather than explicitly declare the dependency and bundle the guava jar
with the deckbuilder plugins, I excised the dependency and wrote a
simple interface layer which presents a collection as if it were a set.
But trust me, it's fine. Probably!

Signed-off-by: Emi <dealuminis@gmail.com>