Add prototype Scryfall search provider
Introduces a dependency on the deckbuilder itself, which is weird given that this is a plugin for that, but... since it's a plugin, it's not technically a dependency cycle. See?

Anyway, this quick-and-dirty implementation forwards your request onto Scryfall and uses the returned cards to filter your search results. Unfortunately, since Scryfall paginates these requests and returns ALL card data, I had to limit the number of results to 10 pages, or 1750 printings.

I've sent an e-mail to Scryfall about a more useful endpoint for this kind of application, but we'll have to wait on that. Meanwhile, this works for relatively precise searches.

Signed-off-by: Emi <dealuminis@gmail.com>