Significant ScryfallApi rewrite to avoid Cloudflare firewall
Scryfall's anti-DDoS service through Cloudflare was flagging Deckbuilder requests. Scryfall's advice was to set an accept header and add a user agent, which I did -- but then realized that I had to make the request in two different places, and that I wasn't handling errors consistently, and so on.

So I rewrote ScryfallApi to handle both JSON and image requests to Scryfall's API, record error responses to the resulting IOException, and properly queue/delay requests to avoid overloading Scryfall's servers.

Signed-off-by: Emi <dealuminis@gmail.com>