Introduction of ScryfallSerde and ScryfallPreferences
Okay, so. My day started with me thinking, "Maybe MessagePack would be faster than gson?"

And it is! But not like, a *ton* faster, and if people still wanna poke around in the data (like me when the data breaks), I'd like to have the option to go back to JSON...

At first, I updated the Deckbuilder's handling of data sources and split the ScryfallDataSource into two different implementations -- but the implementations are inevitably extremely similar. So I instead extracted the serialization and deserialization code into a new class/package emi.lib.mtg.scryfall.serde.ScryfallSerde, and rewrote the data updating and loading methods to be serde-agnostic.

That done, I added a preference using the Deckbuilder's new Preferences.Plugin utility to give folks the option to switch over to MessagePack. I might make that the default in a future release (although by then they'll all be configured to keep Json).

Anyway, it's a huge commit, and I'm sorry about that. Were I more awake, I could probably break it down further. But it all kinda goes together, so here we are!

Signed-off-by: Emi <dealuminis@gmail.com>